pyproject.toml
[tool.poetry]
name = "-"
version = "0.1.0"
description = ""
authors = ["Ilaafanasef21@gmail.com"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.1.6"
python-dotenv = "^1.0.1"
psycopg2-binary = "^2.9.10"
djangorestframework = "^3.15.2"
djangorestframework-simplejwt = "^5.5.0"
stripe = "^11.6.0"
celery = "^5.5.0"
django-celery-beat = "^2.7.0"
requests = "^2.32.3"
django-cors-headers = "^4.7.0"
black = "^25.1.0"
drf-yasg = "^1.21.10"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
